many warnings from cppckeck was fixed (mainly const modifiers) few not initialised variables, few not used removed

+++ MyGUIEngine/include/MyGUI_Macros.h
@@ -26,7 +26,7 @@
 namespace MyGUI
 {
 
-	const size_t ITEM_NONE = ~0;
+	const size_t ITEM_NONE = ~(size_t)0;
 	const int DEFAULT ((int)-1);
 	const float ALPHA_MAX = 1.0f;
 	const float ALPHA_MIN = 0.0f;
